requirements.txt.txt
ultralytics
streamlit
Pillow
numpy
torch
